Adding GIF encoder
- New: GIF animated encoder.

+++ Catalano.Statistics/src/Catalano/Statistics/Tools.java
@@ -245,7 +245,7 @@ public class Tools {
             r += x[i];
         }
         
-        return r / x.length;
+        return r / (double)x.length;
     }
     
     /**
